refactor(ors_control_app): extract routes/regions/lifecycle (sub-folder start)
Begin the routes/regions/ sub-folder per holistic plan decisions. Move the
3 simplest region endpoints into routes/regions/lifecycle.ts:
- GET  /api/regions
- GET  /api/regions/active
- POST /api/regions/active

Add routes/regions/index.ts as the barrel. Future commits will add
catalog.ts, provision.ts, progress.ts as the corresponding handlers move
out of server/index.ts.

Lines down 2,700 -> 2,453.

Phase 1 of holistic-control-app-refactor-final, route extraction step 5.
No behaviour change. tsc passes cleanly.
